Refactor SwiftUI APIs, fix concurrency issues and improve accessibility
- ReadingSettingsView: migrate letterSpacing from Int to Double for 0.5 step support, add .tracking() to preview card, add accessibility labels and selection traits for theme/page-turn buttons
- ReaderContentView: fix AttributedString TTS highlight correctness using NSRange conversion instead of substring search, switch VStack to LazyVStack, replace Color.clear.onTapGesture with Button, simplify ForEach enumerated
- DatabaseService: fix MainActor isolation violation in deleteBook by switching Task.detached to Task { @MainActor in }
- BookDetailView: fix stale book.title reference in delete alert
- TTSService: add @preconcurrency to AVSpeechSynthesizerDelegate for Swift 6 readiness
- LibraryViewModel: move widget data mapping to Task.detached and capture totalBooks before crossing actor boundaries
- Models: add Sendable conformance to Book, Category, Chapter, Bookmark, ReadingSession, ReadingStatistics, BookReadingStat, WidgetBookData, WidgetData
- BookCard: decode PlatformImage in background Task.detached to verify decodeability before caching
- Bulk replace deprecated .cornerRadius() with .clipShape(.rect(cornerRadius:))
- Bulk replace overlay(View) with overlay { } trailing-closure form
- Bulk replace showsIndicators: false with .scrollIndicators(.hidden)
- Replace Task.sleep(nanoseconds:) with Task.sleep(for:)
- LibraryView/ReaderView/ReaderSettingsPanel: replace onTapGesture with Button and add accessibility labels

+++ InkRoom/Sources/Models/ReadingSettings.swift
@@ -3,7 +3,7 @@ import Foundation
 struct ReadingSettings: Codable {
     var fontSize: Int
     var lineSpacing: Int
-    var letterSpacing: Int
+    var letterSpacing: Double
     var readerTheme: ReaderTheme
     var pageTurnStyle: PageTurnStyle
     var ttsRate: Float
